:memo: :: 무한요청해결 및 자동갱신 타이밍 변경
