+++ app/src/main/java/com/g3c1/aide/feature_store/presentation/ui/screen/AddStoreScreen.kt
@@ -121,6 +121,7 @@ fun AddStoreScreen(
                     )
                     addStoreRequest(viewModel, lifecycleCoroutineScope, context) {
                         Toast.makeText(context, "가게 등록에 성공했습니다!", Toast.LENGTH_SHORT).show()
+                        getMyStoresInfoRequest(viewModel, lifecycleCoroutineScope, context)
                         goBackToStoreListPage()
                     }
                 }
@@ -143,7 +144,8 @@ private fun getImageUrlRequest(
                     onSuccess()
                 }
                 is ApiState.Error -> {
-                    Toast.makeText(context, "가게를 등록할 수 없습니다.", Toast.LENGTH_SHORT).show()
+                    Log.d("GetImageUrl", it.status.toString())
+                    Toast.makeText(context, "이미지 용량이 너무 큽니다.", Toast.LENGTH_SHORT).show()
                 }
                 is ApiState.Loading -> {
                 }
@@ -171,6 +173,31 @@ private fun addStoreRequest(
                 }
                 is ApiState.Loading -> {
                 }
+            }
+        }
+    }
+}
+
+private fun getMyStoresInfoRequest(
+    viewModel: StoreViewModel,
+    lifecycleCoroutineScope: LifecycleCoroutineScope,
+    context: Context,
+) {
+    viewModel.getMyStoresInfoRequest()
+    lifecycleCoroutineScope.launch {
+        viewModel.getMyStoresRes.collect { result ->
+            when (result) {
+                is ApiState.Success -> {
+                    Log.d("StoreActivity", result.data.toString())
+                }
+                is ApiState.Error -> {
+                    Log.d("StoreActivity", result.status.toString())
+                    Toast.makeText(
+                        context, "알수 없는 오류가 발생했습니다.", Toast.LENGTH_SHORT
+                    ).show()
+                    viewModel.getMyStoresRes.value = ApiState.Loading()
+                }
+                is ApiState.Loading -> {}
             }
         }
     }
